separacion de archivo monolitico en arquitectura orientada a componentes
